fixup! feat(pathfinder/p2p_network): add tests for GetBlockHeaders
